requirements.txt
fastapi~=0.107.0
langchain~=0.0.352
scikit-learn~=1.3.2
SQLAlchemy~=2.0.23
pandas~=2.1.4
nltk~=3.8.1
uvicorn
google-cloud-aiplatform>=1.38.0
pymysql
python-dotenv